Only update docker daemon config if it exists

+++ Setup/Configure-Windows.ps1
@@ -8,11 +8,13 @@ Set-ItemProperty 'HKLM:\SYSTEM\CurrentControlSet\Control\FileSystem' -Name 'Long
 
 # docker
 $dockerConfigFilePath = "$env:USERPROFILE\.docker\daemon.json"
-$dockerConfig = Get-Content $dockerConfigFilePath | ConvertFrom-Json
-$dockerConfig | Add-Member -MemberType NoteProperty -Name "experimental" -Value $true -Force
-$dockerConfig | Add-Member -MemberType NoteProperty -Name "max-concurrent-downloads" -Value 10 -Force
-$dockerConfig | Add-Member -MemberType NoteProperty -Name "max-concurrent-uploads" -Value 10 -Force
-$dockerConfig | ConvertTo-Json | Out-File $dockerConfigFilePath
+if (Test-Path $dockerConfigFilePath) {
+    $dockerConfig = Get-Content $dockerConfigFilePath | ConvertFrom-Json
+    $dockerConfig | Add-Member -MemberType NoteProperty -Name "experimental" -Value $true -Force
+    $dockerConfig | Add-Member -MemberType NoteProperty -Name "max-concurrent-downloads" -Value 10 -Force
+    $dockerConfig | Add-Member -MemberType NoteProperty -Name "max-concurrent-uploads" -Value 10 -Force
+    $dockerConfig | ConvertTo-Json | Out-File $dockerConfigFilePath
+}
 
 # windows explorer, show file extensions
 Set-ItemProperty -Path 'HKCU:\Software\Microsoft\Windows\CurrentVersion\Explorer\Advanced' -Name HideFileExt -Value 0
